requirements.txt
Flask>=0.12,<1.0
Flask-Cors>=3.0,<3.1
psycopg2>=2.7,<2.8
python-dotenv>=0.8<1.0
